Commit catch up

+++ app/controllers/accounts_controller.rb
@@ -11,19 +11,35 @@ class AccountsController < ApplicationController
 
     def show
         account = Account.find(session[:account_id])
-        render json: account, include: ['assistant', 'assistant.task_agreements', 'assistant.task_agreements.task_category', 'assistant.task_agreements.employer', 'employer', 'employer.task_posts', 'employer.task_posts.task_post_category', 'employer.task_posts.task_category', 'employer.task_agreements', 'employer.task_agreements.task_category', 'employer.reviews.task_post', 'employer.reviews.assistant', 'employer.assistants']
+        render json: account, include: ['assistant', 'assistant.task_agreements', 'assistant.task_agreements.task_category', 'assistant.task_agreements.employer', 'employer', 'employer.task_posts', 'employer.task_posts.task_post_category', 'employer.task_posts.task_category', 'employer.task_agreements', 'employer.task_agreements.task_category', 'employer.reviews.task_post', 'employer.reviews.assistant', 'employer.assistants', 'employer.assistants.account']
     end
 
     def create
-        account = Account.create!(user_params)
+        account = Account.create!(account_params)
         session[:account_id] = account.id
         render json: account, include: ['assistant', 'assistant.task_agreements', 'employer', 'employer.task_posts', 'employer.task_agreements', 'employer.assistants']
     end
 
-    def destroy
+    def update
+        current_user = Account.find(session[:account_id])
         account = Account.find(params[:id])
-        account.destroy
-        head :no_content
+        if current_user.id == account.id
+            account.update!(account_params)
+            render json: account, include: ['assistant', 'assistant.task_agreements', 'employer', 'employer.task_posts', 'employer.task_agreements', 'employer.assistants']
+        else
+            render json: { error: ["Sorry, you're not authorized"] }, status: :unauthorized
+        end
+    end
+
+    def destroy
+        current_user = Account.find(session[:account_id])
+        account = Account.find(params[:id])
+        if current_user.id == account.id
+            account.destroy
+            head :no_content
+        else
+            render json: { error: ["Sorry, you're not authorized"] }, status: :unauthorized
+        end
     end
 
     private
@@ -36,8 +52,12 @@ class AccountsController < ApplicationController
         return render json: { error: "Not authorized" }, status: :unauthorized unless session.include? :account_id
     end
 
-    def user_params
+    def account_params
         params.permit(:name, :address, :phone, :email, :password, :password_confirmation, :image)
     end
+
+    # def current_user
+    #     @current_user ||= Account.find(session[:account_id])
+    # end
 
 end
