+++ app/controllers/task_agreements_controller.rb
@@ -1,4 +1,9 @@
 class TaskAgreementsController < ApplicationController
+    rescue_from ActiveRecord::RecordInvalid, with: :render_unprocessable_entity
+    rescue_from ActiveRecord::RecordNotFound, with: :render_not_found_response
+    before_action :authorize
+    skip_before_action :authorize, only: [:index, :show]
+
 
     def index
         task_agreements = TaskAgreement.all
@@ -6,8 +11,9 @@ class TaskAgreementsController < ApplicationController
     end
 
     def create
-        task_agreement = TaskAgreement.create!(task_agreement_params)
-        render json: task_agreement
+        current_user = Account.find(session[:account_id])
+        task_agreement = current_user.employer.task_agreements.create!(task_agreement_params)
+        render json: task_agreement, include: ['assistant', 'assistant.account']
     end
 
     def show
@@ -16,22 +22,40 @@ class TaskAgreementsController < ApplicationController
     end
 
     def update
-        task_agreement = TaskAgreement.find(params[:id])
-        task_agreement.update(
-            task_agreement_notes: params[:task_agreement_notes],
-            hourly_rate: params[:hourly_rate],
-            is_completed: params[:is_completed]
-        )
-        render json: task_agreement
+        current_user = Account.find(session[:account_id])
+        task_agreement = current_user.employer.task_agreements.find(params[:id])
+        if task_agreement.id == params[:id].to_i
+            task_agreement.update!(task_agreement_params)
+            render json: task_agreement
+        else
+            render json: { error: ["Sorry, you're not authorized"] }, status: :unauthorized
+        end
     end
 
     def destroy
-        task_agreement = TaskAgreement.find(params[:id])
-        task_agreement.destroy
-        head :no_content
+        current_user = Account.find(session[:account_id])
+        task_agreement = current_user.employer.task_agreements.find(params[:id])
+        if task_agreement.id == params[:id].to_i
+            task_agreement.destroy
+            head :no_content
+        else
+            render json: { error: ["Sorry, you're not authorized"] }, status: :unauthorized
+        end
     end
 
     private
+
+    def render_unprocessable_entity(invalid)
+        render json: { error: invalid.record.errors.full_messages }, status: :unprocessable_entity
+    end
+
+    def render_not_found_response(error)
+        render json: { error: [error.message] }, status: :not_found
+    end
+
+    def authorize
+        return render json: { error: "Not authorized" }, status: :unauthorized unless session.include? :account_id
+    end
 
     def task_agreement_params
         params.permit(:assistant_id, :employer_id, :hourly_rate, :is_completed, :task_agreement_notes, :task_post_id)
